multiagent wrapper
