+++ highway_env/envs/common/abstract.py
@@ -393,11 +393,3 @@ class AbstractEnv(gym.Env):
             else:
                 setattr(result, k, None)
         return result
-
-
-class MultiAgentWrapper(Wrapper):
-    def step(self, action):
-        obs, reward, done, info = super().step(action)
-        reward = info["agents_rewards"]
-        done = info["agents_dones"]
-        return obs, reward, done, info
